refactor: query-key, end-point 위치 변경
